feat: Resize window to image dimensions
The implementation is not perfect.

When a large image is loaded, a blur artifact appears, that disppears
as soon as the image is re-rendered. (By resizing the window once
more for example.)

That visual artifact doesn't appear for all heights and widths,
but consistently appears when reizing the window to the dimensions
of the image, if the image is large enough.

The size of the artifact doesn't seem to depends on the dimensions
of the image, aside from being or not visible.